refactor(backend): create multiple piece metadata services for dev and prod
